Fix: Move classes
