+++ src/FieldResolution/ResolutionStrategy.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * @see https://github.com/ergebnis/factory-bot
  */
 
-namespace Ergebnis\FactoryBot\Strategy;
+namespace Ergebnis\FactoryBot\FieldResolution;
 
 use Ergebnis\FactoryBot\Count;
 use Ergebnis\FactoryBot\FieldDefinition;
@@ -21,27 +21,19 @@ use Faker\Generator;
 /**
  * @internal
  */
-final class WithoutOptionalStrategy implements ResolutionStrategy
+interface ResolutionStrategy
 {
+    /**
+     * @return mixed
+     */
     public function resolveFieldValue(
         Generator $faker,
         FixtureFactory $fixtureFactory,
         FieldDefinition\Resolvable $fieldDefinition,
-    ) {
-        if ($fieldDefinition instanceof FieldDefinition\Optional) {
-            return null;
-        }
-
-        return $fieldDefinition->resolve(
-            $faker,
-            $fixtureFactory,
-        );
-    }
+    );
 
     public function resolveCount(
         Generator $faker,
         Count $count,
-    ): int {
-        return $count->minimum();
-    }
+    ): int;
 }
