+++ test/Integration/FixtureFactoryTest.php
@@ -16,8 +16,8 @@ namespace Ergebnis\FactoryBot\Test\Integration;
 use Ergebnis\FactoryBot\Count;
 use Ergebnis\FactoryBot\EntityDefinition;
 use Ergebnis\FactoryBot\FieldDefinition;
+use Ergebnis\FactoryBot\FieldResolution;
 use Ergebnis\FactoryBot\FixtureFactory;
-use Ergebnis\FactoryBot\Strategy;
 use Example\Entity;
 use Faker\Generator;
 use PHPUnit\Framework;
@@ -30,7 +30,7 @@ use PHPUnit\Framework;
 #[Framework\Attributes\UsesClass(FieldDefinition\Reference::class)]
 #[Framework\Attributes\UsesClass(FieldDefinition\Sequence::class)]
 #[Framework\Attributes\UsesClass(FieldDefinition\Value::class)]
-#[Framework\Attributes\UsesClass(Strategy\DefaultStrategy::class)]
+#[Framework\Attributes\UsesClass(FieldResolution\DefaultStrategy::class)]
 final class FixtureFactoryTest extends AbstractTestCase
 {
     public function testCreateOneDoesNotPersistEntityByDefault(): void
