+++ test/Unit/FieldResolution/DefaultStrategyTest.php
@@ -11,22 +11,22 @@ declare(strict_types=1);
  * @see https://github.com/ergebnis/factory-bot
  */
 
-namespace Ergebnis\FactoryBot\Test\Unit\Strategy;
+namespace Ergebnis\FactoryBot\Test\Unit\FieldResolution;
 
 use Ergebnis\FactoryBot\Count;
 use Ergebnis\FactoryBot\FieldDefinition;
+use Ergebnis\FactoryBot\FieldResolution;
 use Ergebnis\FactoryBot\FixtureFactory;
-use Ergebnis\FactoryBot\Strategy;
 use Ergebnis\FactoryBot\Test;
 use PHPUnit\Framework;
 
-#[Framework\Attributes\CoversClass(Strategy\WithOptionalStrategy::class)]
+#[Framework\Attributes\CoversClass(FieldResolution\DefaultStrategy::class)]
 #[Framework\Attributes\UsesClass(Count::class)]
 #[Framework\Attributes\UsesClass(FieldDefinition::class)]
 #[Framework\Attributes\UsesClass(FieldDefinition\Optional::class)]
 #[Framework\Attributes\UsesClass(FieldDefinition\Value::class)]
 #[Framework\Attributes\UsesClass(FixtureFactory::class)]
-final class WithOptionalStrategyTest extends Test\Unit\AbstractTestCase
+final class DefaultStrategyTest extends Test\Unit\AbstractTestCase
 {
     public function testResolveFieldValueResolvesOptionalFieldDefinitionWithFakerAndFixtureFactoryWhenFakerReturnsTrue(): void
     {
@@ -39,7 +39,7 @@ final class WithOptionalStrategyTest extends Test\Unit\AbstractTestCase
 
         $fieldDefinition = FieldDefinition::optionalValue($faker->sentence());
 
-        $strategy = new Strategy\WithOptionalStrategy();
+        $strategy = new FieldResolution\DefaultStrategy();
 
         $resolved = $strategy->resolveFieldValue(
             new Test\Double\Faker\TrueGenerator(),
@@ -55,7 +55,7 @@ final class WithOptionalStrategyTest extends Test\Unit\AbstractTestCase
         self::assertSame($expected, $resolved);
     }
 
-    public function testResolveFieldValueResolvesOptionalFieldDefinitionWithFakerAndFixtureFactoryWhenFakerReturnsFalse(): void
+    public function testResolveFieldValueResolvesOptionalFieldDefinitionToNullWhenFakerReturnsFalse(): void
     {
         $faker = self::faker();
 
@@ -66,7 +66,7 @@ final class WithOptionalStrategyTest extends Test\Unit\AbstractTestCase
 
         $fieldDefinition = FieldDefinition::optionalValue($faker->sentence());
 
-        $strategy = new Strategy\WithOptionalStrategy();
+        $strategy = new FieldResolution\DefaultStrategy();
 
         $resolved = $strategy->resolveFieldValue(
             new Test\Double\Faker\FalseGenerator(),
@@ -74,12 +74,7 @@ final class WithOptionalStrategyTest extends Test\Unit\AbstractTestCase
             $fieldDefinition,
         );
 
-        $expected = $fieldDefinition->resolve(
-            $faker,
-            $fixtureFactory,
-        );
-
-        self::assertSame($expected, $resolved);
+        self::assertNull($resolved);
     }
 
     public function testResolveFieldValueResolvesFieldDefinitionWithFakerAndFixtureFactoryWhenFakerReturnsTrue(): void
@@ -93,7 +88,7 @@ final class WithOptionalStrategyTest extends Test\Unit\AbstractTestCase
 
         $fieldDefinition = FieldDefinition::value($faker->sentence());
 
-        $strategy = new Strategy\WithOptionalStrategy();
+        $strategy = new FieldResolution\DefaultStrategy();
 
         $resolved = $strategy->resolveFieldValue(
             new Test\Double\Faker\TrueGenerator(),
@@ -120,7 +115,7 @@ final class WithOptionalStrategyTest extends Test\Unit\AbstractTestCase
 
         $fieldDefinition = FieldDefinition::value($faker->sentence());
 
-        $strategy = new Strategy\WithOptionalStrategy();
+        $strategy = new FieldResolution\DefaultStrategy();
 
         $resolved = $strategy->resolveFieldValue(
             new Test\Double\Faker\FalseGenerator(),
@@ -139,12 +134,10 @@ final class WithOptionalStrategyTest extends Test\Unit\AbstractTestCase
     #[Framework\Attributes\DataProviderExternal(Test\DataProvider\IntProvider::class, 'greaterThanOrEqualToZero')]
     public function testResolveCountResolvesCountWithFakerWhenCountIsExact(int $value): void
     {
-        $faker = self::faker();
-
-        $strategy = new Strategy\WithOptionalStrategy();
+        $strategy = new FieldResolution\DefaultStrategy();
 
         $resolved = $strategy->resolveCount(
-            $faker,
+            self::faker(),
             Count::exact($value),
         );
 
@@ -156,7 +149,7 @@ final class WithOptionalStrategyTest extends Test\Unit\AbstractTestCase
     {
         $maximum = self::faker()->numberBetween($minimum + 1);
 
-        $strategy = new Strategy\WithOptionalStrategy();
+        $strategy = new FieldResolution\DefaultStrategy();
 
         $resolved = $strategy->resolveCount(
             new Test\Double\Faker\MinimumGenerator(),
@@ -166,12 +159,7 @@ final class WithOptionalStrategyTest extends Test\Unit\AbstractTestCase
             ),
         );
 
-        $minimumGreaterThanZero = \max(
-            1,
-            $minimum,
-        );
-
-        self::assertSame($minimumGreaterThanZero, $resolved);
+        self::assertSame($minimum, $resolved);
     }
 
     #[Framework\Attributes\DataProviderExternal(Test\DataProvider\IntProvider::class, 'greaterThanZero')]
@@ -179,7 +167,7 @@ final class WithOptionalStrategyTest extends Test\Unit\AbstractTestCase
     {
         $minimum = self::faker()->numberBetween(0, $maximum - 1);
 
-        $strategy = new Strategy\WithOptionalStrategy();
+        $strategy = new FieldResolution\DefaultStrategy();
 
         $resolved = $strategy->resolveCount(
             new Test\Double\Faker\MaximumGenerator(),
@@ -199,7 +187,7 @@ final class WithOptionalStrategyTest extends Test\Unit\AbstractTestCase
         $minimum = $faker->numberBetween(1);
         $maximum = $faker->numberBetween($minimum + 1);
 
-        $strategy = new Strategy\WithOptionalStrategy();
+        $strategy = new FieldResolution\DefaultStrategy();
 
         $resolved = $strategy->resolveCount(
             $faker,
@@ -209,12 +197,7 @@ final class WithOptionalStrategyTest extends Test\Unit\AbstractTestCase
             ),
         );
 
-        $minimumGreaterThanZero = \max(
-            1,
-            $minimum,
-        );
-
-        self::assertGreaterThanOrEqual($minimumGreaterThanZero, $resolved);
+        self::assertGreaterThanOrEqual($minimum, $resolved);
         self::assertLessThanOrEqual($maximum, $resolved);
     }
 }
